fix(status): restore StatusWriter tests; new tests in test_status_command.py
Task 2 overwrote the pre-existing carta/tests/test_status.py (embed
StatusWriter unit tests). Restore it and move the new status-command tests
to test_status_command.py to avoid the filename collision.
